+++ frontend/src/components/editing/QandAArea.tsx
